Refactor admin models for organization django-app
