+++ kobo/apps/organizations/admin/__init__.py
@@ -1,0 +1,4 @@
+from .organization import OrgAdmin
+from .organization_owner import OrgOwnerAdmin
+from .organization_invite import OrgInvitationAdmin
+from .organization_user import OrgUserAdmin
