+++ kobo/apps/organizations/admin/organization_invite.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from ..models import OrganizationInvitation
+
+
+@admin.register(OrganizationInvitation)
+class OrgInvitationAdmin(admin.ModelAdmin):
+    pass
